+++ src/content/state.ts
@@ -46,18 +46,4 @@ export function getState() {
   };
 }
 
-// Calculate the playback rate needed to achieve the target frequency
-export function calculatePlaybackRate(): number {
-  return _targetFrequency / getReferenceFreq(_targetFrequency);
-}
 
-// Check if given playback rate comes from a non-standard frequency and should be reset to 1 in pitch mode
-export function shouldResetPlaybackRate(mediaPlaybackRate: number): boolean {
-  const nonStandardFrequencies: Frequency[] = [432, 528];
-  const expectedRates = nonStandardFrequencies.map(
-    (freq) => freq / getReferenceFreq(freq)
-  );
-  return expectedRates.some(
-    (rate) => Math.abs(mediaPlaybackRate - rate) < 0.0001
-  );
-}