+++ src/shared/constants.ts
@@ -1,6 +1,5 @@
 import { GlobalState } from "./types";
 
-export const WORKLET_PATH = chrome.runtime.getURL("soundtouch-worklet.js");
 export const C5_STANDARD_FREQUENCY = 523.2511306011972;
 export const A4_STANDARD_FREQUENCY = 440;